+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - more options for customizing tree sidebar
 - describe various options in the readme
 - fix sidebar right resizing
+- improve sidebar right handling
 
 ### v0.5.0, Sep 04, 2014
 - key bindings work from the sidebar now too